exec/expand_path: Copy absolute/relative path with ft_strdup
This is not strictly necessary, but fixes crash when cmd_path is
freed at the end of executor(). Previously it was not malloced with
absolute and relative paths, i.e. when the command is not searched
from $PATH. The ft_strdup mallocs the path string, fixing the crash
when attempting too free the non-malloced string.

Fixes: #2