exec: Do not ignore zero length commands
Now we print permission denied error instead (which is what zsh does
as well). The downside is that we get the same error from executing
a zero length variable, but this that is more of a special case so
this should be better.

See https://github.com/linhtng/minishell/issues/17